updated exec and container files chore
