+++ biosimulator_processes/container-assets/Dockerfile-base
@@ -4,7 +4,7 @@ ENV DEBIAN_FRONTEND=noninteractive
 
 WORKDIR /app
 
-# COPY . /app
+COPY ./biosimulator_processes /app
 
 # COPY ./scripts /app
